requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
python-dotenv
openai

# testing
pytest
httpx

